Add the option to buy an island.
I was in an Internet-based chat-group and I introduced them
to this little game of spending (billionaires') money.
I heard a desire to buy an island "for us to party", but
did not manage to find an island among the existing options.
Another chat-ter said: "you have all this money but cannot buy
an island". Well, I wrote the code/script(s) to enable me
to buy an island in this little game.